schedule code tv channel
This code was taken from the town website. It contains:
A) a table of the daily programs with a number  of fields Day,
date, time,
1)program, length, next air date, watching,
2) uses a calendar interface
3) controls forward and backward month tool
4) also controls day of the month
5) returns the a new  page based on the calendar interface tool
selections on that webpage.

the fields in the table could be used as links such as 
watch.... could  open a viewer window, play that mov file with scroll controls and time code at the same time, offer info specific to the program author, bio, series name, first aired, link to local channel (not ours) to see what else they offer, simular programs ect.(closed captioning LOL) or a simple text reader from the audio track. (speech to text)
the entire data base of all tv programs on the server could also be searched based on a key word,title, subject
this is a starting point